Make makeSimpleDestructiveAssignment condition parameter non-nullable
If condition is truth constant, it has the same behavior as null had
previously. We can rely on the SMT solver to replace if-then-else with
constant condition into just the always taken case.

The performance impact is negligible as the only place where null was
used is zero-initialization memory allocation (calloc).

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/predicate-memory-functions@43514 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c