Rework pointer-equivalent types and conditions in assignments
In SliceExpression.getDummyResolvedExpression(), I removed the
pointer-equivalent integer type parameter as we can supply arbitrary
integer type for zero-index construction instead (int chosen as it is
the default). This means it is no longer necessary to store the type in
AssignmentHandler.

In MemoryManipulationFunctionHandler, the type is used only for
CExpression operations and can reasonably be the normal pointer-sized
type.

In AssignmentQuantifierHandler, it is used for construction of slice
conditions and encoded quantifier variables. I have therefore chosen the
unsigned type which can contain all pointer values, made comparisons
unsigned, and removed the 0 <= i constraints as they are no longer
necessary.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/predicate-memory-functions@43515 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c