Change quantifier handling to work on each partial assignment separately
The previous way of unrolling/encoding for all simple partial
assignments at once was an extreme performance bottleneck when assigning
types with a large amount of fields as there could be tens of indices
that were orthogonal to each other. They were unrolled at once, causing
an exponential slowdown.

Now, indices for each partial assignment are unrolled/encoded
separately, so the problem is eliminated.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/predicate-memory-functions@43524 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c