Rewrite memset byte repeating to be handled inside assignment handlers
The byte repeating inside MemoryManipulationFunctionHandler did not play
well with floats as it required an integer CType with equivalent size,
which may not be always available (especially for long double).

I rewrote the byte repeating to be selected by a new
AssignmentOptions.ConversionType enum value, which resolves this problem
at the cost of some complications to AssignmentFormulaHandler (which
handles the repeating) and AssignmentQuantifierHandler (which handles
all-zeros or all-ones check if bitfields are to be set).

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/predicate-memory-functions@43526 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c