Remove option to force memory assignment options to encode quantifiers
This option is not useful for anyone else than people working on
quantifier encoding support in the future, and those will probably need
a more fine-grained system of options for their internal use. I am
removing the option in the meantime. The possibility to force encoding
for some assignments only remains in AssignmentOptions for future use.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/predicate-memory-functions@43580 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c