Resolve problematic SMTHeap contract by adding SMTMultipleAssignmentHeap
We need multiple-assignment heaps to serve as backing heaps to
SMTHeapWithByteArray. At the same time, SMTHeapWithByteArray would be
more complex if multiple assignments were to be supported there, which
is not required. So I split the multiple-assignment method to a new
interface to resolve this problem in a type-conscious way.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/predicate-memory-functions@43582 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c