Improve the handling of values of string literals
So far there were the following methods in CStringLiteralExpression:
- getValue: Returns the string literal including quotation marks.
- toASTString: Same as getValue.
- getContentString: Returns the string content (without quotation marks)
  but misses the null terminator that is included according to the
  C standard.

None of these returned the actual content of the string literal
according to the C standard.

Now we provide:
- toASTString: Returns the string literal including quotation marks.
- getContentWithNullTerminator: Returns the content according to the
  C standard (no quotation marks, with null terminator).
- getContentWithoutNullTerminator: Same as getContentWithNullTerminator
  but without the null terminator (useful for some callers).
- getValue: marked as deprecated

This should make it easier for callers to pick the right method.
Existing callers are adjusted,
for most the newly called method has the same semantics as the one that
they called until now, but a few are changed such that they now use the
content of the string instead of the syntactic representation.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@43700 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c