Fix type and sizeof for string literals
The type of a string literal in our internal representation
is char* (produced by Eclipse CDT).
This means that code that evaluates sizeof on string literals
will produce 4 or 8, but not the actual size.
We could fix this by fixing each sizeof implementation,
but there are several of these.
And in fact § 6.4.5 (6) of C11 declares string literals
to be an array of type char, so it should be possible to fix the type of
string literals to always be char[] with the correct length.
Then all sizeof calculations will automatically do the right thing.
We have to adjust a few pieces of code that did not expect
expressions to have array types, though.

Fixes #1087

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@43703 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c