Fix array-to-pointer assignment dereferencing where it should not
The AssignmentFormulaHandler.makeSimpleDestructiveAssignment contains a
branch based on pRvalueType being an array type, in which case it
doesn't dereference the aliased location which is already an
__ADDRESS_OF, but in the slice assignments, the pRvalueType never was an
array type when assigning to a pointer. I have modified the behaviour so
the array type is supplied in pRvalueType to
makeSimpleDestructiveAssignment in the special case of array-to-pointer
assignment.

This should fix
test/programs/simple/pointer_aliasing/array_to_pointer_assignment.i
test/programs/simple/assignment-pointer-array.c

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/predicate-memory-functions@43724 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c