Unify AssignmentFormulaHandler and AddressHandler getValueFormula(...)
The functions have done basically the same thing with some differences,
mainly the AssignmentFormulaHandler one returning an Optional which was
empty when the variable was nondet, and AddressHandler one making a new
variable; and the AddressHandler supporting safe dereference flag. They
are now unified in AddressHandler, with no duplicate code.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/predicate-memory-functions@43725 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c