Add warning when memory manipulation fn definitely loses soundness
Current implementation of the memory manipulation function treats all
array elements (given by the destination/source pointer) as a whole, so
if the byte size given as a parameter is not perfectly divisible by
element size, there is loss of soundness. In case of literals, we now
catch that and print a warning.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/predicate-memory-functions@43761 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c