Refactor combination of partial spans in AssignmentFormulaHandler
The current code expands the formula for each span to the full length
with zeroes and then computes the bitwise or of all these formulas.
The new code just concatenates the formulas for each of the spans.

This is hopefully easier to understand both when looking at the code
as well as when looking at the formulas,
and likely easier for the SMT solver as well.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/predicate-memory-functions@43763 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c