Fix bug from r43763 / 06434d4: bv concatenation was in wrong order
concat() expects the most significant bits as first parameter and
the least significant bits as second parameter.
But partialRhsMap is ascending and starts with least significant bits,
so a reduction on it needs to swap the arguments for the concat() call.

Also fix the parameter names of concat() to make this clear.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/predicate-memory-functions@43766 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c