Fix string literal overapproximation to be assigned as previously
The default handling of string literals with
handleStringLiteralInitializers=false is to assign them to an
uninterpreted function (__string__ index), where index is a number, so
that two character arrays that were declared with the same literal
initializer are treated as having the same elements.

Reworking string literals to have array type forced assignment by
repeated literal reads and array writes, which is a bit less
overapproximating, but produces large formulas, doing away with the
reason for optimization in the first place. I have fixed this by
restoring the previous behaviour in the case of string literals with
handleStringLiteralInitializers=false. This also does away with warnings
for skipping assignment due to different literal and declared array
sizes.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/predicate-memory-functions@43767 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c