Cherry-pick r43712 / 6ab58e29 from branch predicate-memory-functions
In the work for #664 / !107 it was noticed that our optimization
for irrelevant assignments in the predicate analysis
had the unwanted side effect of ignoring all calls to extern functions
if they appear inside an assignment.
This also affected the deny listing of extern functions such as memset
from r32980 / 80e35da7.

This means that so far the statement "memset(...);"
would trigger an error about memset being unsupported,
but the statement "p = memset(...);" would be silently ignored
if p does not happen to be relevant for the property.
That doesn't really make sense,
because the side effect of the function could of course
still be relevant even if the resulting pointer is not.
Furthermore, the same change was made to the InvariantsCPA
in r38844 / faef6e11 in order to make it consistent with the PredicateCPA,
but there the check never considered relevancy of assignments,
so the InvariantsCPA was unknowingly made stricter than the PredicateCPA.

Now with this change, calls to functions such as memset
are always considered relevant in the PredicateCPA,
and thus will always trigger the error about unsupported functions,
i.e., it has the same behavior as the InvariantsCPA.

On the SV-Benchmarks set, this will lead to about 100 more errors
and (interestingly) no less wrong results,
likely because most of the memset calls are irrelevant.
But it is still conceptually better to have this check
that ensures soundness.

Users who want to (unsoundly) ignore such memory-manipulation functions
can continue doing so by using the option
cpa.predicate.allowedUnsupportedFunctions.

And we will have a working implementation of memset etc.
thanks to the work in !107 by Jan Onderka soon.



git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@43787 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c