Disable handling of memory-manipulation functions for now.
This branch contains both a rewrite of the assignment handling
of the PredicateCPA as well as support for these functions.
We want to keep refactorings and new features separate
in order to make understanding test results, debugging, bisecting, etc. easier.
So it was decide to disable the handling of memory functions,
then merge this branch, and then later on enable the functions.
A detailed discussion is in
https://gitlab.com/sosy-lab/software/cpachecker/-/merge_requests/107#note_1410639009

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/predicate-memory-functions@43790 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c