Replace option cpa.predicate.ignoreExtractConcat with ignoreExtractExtend
This option is only relevant for the predicate analysis
with linear approximation, and on by default.

First, the old name (and documentation) was bad
because it affected not only extract and concat, but also extend.
The option was introduced in r7386 / 7a74ae43
and handling of extend was added in r9274 / 0114aec4.

Second, while there are arguments for replacing extract and extend
with noops, this is much more questionable for concat,
because concat has two arguments, not just one.
So with this option, the analysis implemented "concat(a,b)" as "b"!

Now we remove the handling of concat from this option
such that concat is always implemented as a UF.
And we fix the name such that it correctly lists the affected
operations.

More discussion of this is in
https://gitlab.com/sosy-lab/software/cpachecker/-/issues/1101#note_1435648177

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@43898 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c