Avoid using Set as type for set entries or map keys
Using mutable objects as set entries or map keys is error-prone,
as the set/map will exhibit undefined behavior
if the object is (accidentally) changed while it is in the set/map.
So it is better to use ImmutableSet.

In some of the cases here this is only a type change
and all sets were already immutable,
but in others there were actually mutable sets used,
in one case even instances that were passed in from the outside,
so other code could easily change them.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@43905 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c