Fix a case of accessing a field before it is assigned
This is a tricky situation:
The ShutdownRequestListener may be called before the invariant generator
is created, and if it accesses the invariantGenerator field,
it may read a null value.
This was handled in r19388 / 72470b1e,
but I think the code is even more problematic.
Typically the ShutdownRequestListener is called from a different thread,
and accessing final fields is fine, but only after the constructor has
finished. This is not guaranteed here.

So it is better to create the ShutdownListener later
and let it bind to the actual invariant generator value
instead of having it access the invariantGenerator field.
This is guaranteed to be thread-safe and avoids the null value.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@43928 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c