Add config option for reproduction Artefact file
Options specified in this properties file are applied on top of the normal cpachecker config file. This can be used to apply statistics limits on a task-wise basis for a big set of task files. In Benchexec, one can dynamically specify artefacts files of a previous run with the  variable.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/statistics-refactor@43939 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c