Fix crash in termination analysis if used without -noout
We disable writing of the standard violation witnesses
in the termination config, because they are useless.
However, the witnesses were still internally created
except if -noout was used, and this crashed.
Now we generate the witness only if we also write it do disk.
This also has the chance to make all configurations a little bit faster
if they find a violation and no witness should be written do disk.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@43944 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c