Call makeVariable() lazily in one place in predicate analysis
This has no effect on the produced SMT formulas,
but it seems to have some effect on the internal state of the SMT solver,
and this has lead to different models being produced,
in particular for eca-rers2012/Problem01_* and -terminationAnalysis.
The previous results were better, so we revert this change
from r43218 / ece862c4.

Fixes #1102.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@43950 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c