Rewrite branching in equals methods into single return (6)
Here we apply this to cases with a single branch and an else branch
for the "return false;" case.
This also occurs more often outside of equals(),
and we rewrite this as well, as the new code is just simpler.

Part of #1109.

Commit created with
perl -0777 -i -pe 's/if \(([^ ();]*) instanceof ([^ ();]*)\) {\n *return ([^;]*);\n *} else {(\n *)*return false;\n *}/return \1 instanceof \2 && \3;/' src/**/*.java

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/1109-refactor-equals-methods@43994 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c