Rewrite equals/hashCode of counterexample.LeftHandSide
If a super class has fields, it makes more sense to check for equality
of these fields in the super class.
Otherwise we have to duplicate this logic in all sub classes.
Same for hashCode().
And we do not need getClass() checks,
regular instanceof checks are sufficient for this hierarchy.

At the same time lets mark the hierarchy as sealed.

Part of #1109.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/1109-refactor-equals-methods@44001 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c