Restructure three last equals() implementations.
These all share the pattern that depending on some condition,
one of two different equality checks needs to be made.
This is too complex to force it into a single expression.
But it is still a good idea to have a common pattern for this case
that makes this explicit and is similar to our standard pattern
for the remainder of the checks.

Part of #1109.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/1109-refactor-equals-methods@44002 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c