Add comments to some equals methods with more complex checks
This includes cases for example where we need a disjunction.
The main goal of having the comment is for a future CI check
to be able to distinguish between cases where a non-standard equals
implementation is really necessary and cases where the developer just
did not know the standard pattern.

Part of #1109.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/1109-refactor-equals-methods@44004 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c