Add documentation and CI check for our standard patterns for equals()
The CI check is implemented as a bash script that requires grep.
This is not so nice for our Windows users,
but unfortunately <containsregexp> of Ant matches only per line,
and we need regexp matches that span lines.
An alternative would be a Python script,
but I don't think this would help much.

For this reason the new check is also not part of
"ant standard-checks".

Fixes #1109.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/1109-refactor-equals-methods@44005 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c