Improve documentation of InvariantExportAlgorithm
The description was burried in a commit message. Better have this as
Javadoc to the class. When I want to know what a class is good for,
I usually expect this to be written in the Javadoc. After all,
commit messages are not easily visible, e.g. if the file is reformatted
one does not even see this in a blame view.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@44051 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c