Add way to output proof witness in new yaml format
This works by setting cpa.arg.yamlProofWitness=<filename>
Once this option is set, invariants will be extracted from the
Witness object that is generated in the ARGStatistics and
dumped into the new yml format. For now, we just dump the
loop invariants, i.e., those that belong to a node that is
a loop head. In the future we might also want to dump
arbitrary location invariants.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@44054 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c