Disable simplification of symbolic values in symbolic execution
Over all SV-COMP runs there are 0 occurrences where this simplification worked,
but this cost over 90% of the symbolic execution's runtime.
Without this option, symbolic execution is more than 4 times faster on considered examples.

We still have to see whether to fully remove the ConstraintsStrengthenOperator or whether
there is a bug in the implementation.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@44100 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c