Add memory function pointer type short-circuit if byte heap is used
This change should make byte array heap perform better when memory
functions are used, especially not rejecting tests where the actual type
of source / destination cannot be inferred. It should also be more sound
(fractional element set is resolved correctly as exactly the bytes that
should be manipulated are manipulated).

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/faster-byte-heap@44147 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c