Do not convert to progenitor if using byte array heap
Conversion to progenitor is only necessary to soundly resolve simple
cases of unions when different heaps are used. Byte array heap already
soundly resolves unions by the virtue of writing to and reading from the
same heap elements, so this is unnecessary.

Not converting to progenitor may improve the speed of checking as there
will not be duplicate assignments to the same actual byte array heap
element due to a union in progenitor.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/faster-byte-heap@44148 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c