Add a blank edge between condition and loop body for do-while loops
This is part of #1113 and the detailed description is there.
The summary is that whenever we might want to re-add a state to the
waitlist in the analysis, it is better if that state's location has only
one outgoing edge. And parents of loop heads are likely to be re-added
(e.g., in incremental BMC).
So while this additional blank edge is not strictly necessary for
correct CFA semantics, it ensures a parent with a single outgoing edge
for loop-heads of do-while loops.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@44160 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c