Restrict search for counterexample in BMC to relevant subgraph of ARG
When we reconstruct the path to the violating target state from the model,
we do not need to consider the full ARG, but can restrict ourselves
to those states on the paths to target states.
This is the same what we do in predicate abstraction as well.
It could be more efficient (less calls to model evaluation to the solver)
and it can avoid some cases of crashes due to invalid ARG branching.

Fixes #1113.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@44162 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c