Fix wrong proofs by InvariantsCPA due to wrong bitvector arithmetic
The CPA used division in order to resolve equations with
multiplications, but for bitvectors division is not the inverse of
multiplication (due to overflows and truncation).
So this could lead to wrong results.
We have to be more imprecise and assume that something like
"a * b = c" is satisfiable even if a and c are known and just b is
unknown.

Fixes #1114.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@44174 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c