Use a single data structure to store branch/merge node mappings
We only care about branchings with a clear correspondence between
a branching point and a merging point, so there is no need to use two
separate maps.
If we use a single data structure, we have the additional advantage that
we avoid computing merge/branch node mappings if we've already computed
branch/merge node mappings (same for the reverse case).

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/1065-stack-overflow-loop-free-section-finder@44358 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c