Add proper option to enable yaml witnesses
We will add support for violation witnesses, so we cannot
base the decision whether or not to create InvariantWitnessWriter
off of the existence of a yaml proof witness path.
Rather introduce a separate option to safeguard this new feature,
which defaults to false, i.e., nobody runs the new code unless
they explicitly set that option to true.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@44379 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c