Improve yaml witness WaypointRecord class
We use actual enums instead of plain strings now. There is also a special enum value
UNKNOWN for those cases where we do not recognize the keyword. This way, if someone
extends the format and we get a new key, we will already have handled that unknown
case (we will most-likely want to ignore waypoints of unknown type or action).

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@44382 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c