Add support for function_enter waypoint records
Currently, we only need this for the reach_error call,
such that we can convert it to a target waypoint at the end
of the violation sequence, but later we might want to have
these function enter nodes also inbetween in the waypoint sequence.
For now, we filter them out, since their semantics is not clear yet
in case a function call appears in the middle of a statement, not
as a standalone statement.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@44385 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c