Avoid uninstantiating an instantiated path formula
In the encoding of IMC/ISMC, the initial (prefix) path formula may contain variables that are instantiated by different SSA indices.
Therefore, we cannot simply uninstantiate the formula and store it in the fixed point.
One way to work around this is by exisitentially quantifying those variables.
However, MathSAT 5 does not support quantifier elimination,
so we here use interpolation to compute approximation.
(related to #997)

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/export-btor2c-invariants@44402 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c