Refactor GraphToPixelsWriter to use own Options class and initialize lazily
As a result, java.awt (used in GraphToPixelsWriter through Apache Batik) is loaded only when needed.
It's always a good idea to load a large package like java.awt only when
actually used,
but this also allows us to use GraalVM native-image with CPAchecker, which does not support java.awt yet.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/native-image@44484 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c