Fix Optional.of -> Optional.ofNullable
When reading in a yml witness the field will have value
null in case it is not present in the witnes. Ideally
we catch this during parsing of the witness and reject it,
but for now we must expect null to be a possible value.
Using Optional.of will always lead to an optinal with
a value, which is not what we want.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@44490 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c