Fix bug regarding bitwise negation in assumptions in InvariantsCPA
Fixes #1130

If bitwise negation occurs in an assumption of the form "x == ~y",
the InvariantsCPA attempted to check satisfiability of this assumption
by checking whether the set of values of y overlaps with the complement
of values of x.
Unfortunately, bitwise negation as an operation has nothing to do with
interval complement, so this was unsound.
Now we remove this check and overapproximate.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@44498 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c