Rename various CSimpleType methods like isSigned
This method does not return whether the type is signed,
it returns whether the represented type has an explicit "signed"
modifier, but there are signed types without such a modifier
(like "int").
Misuse of this method causes bugs like #1137.
In fact, this method should rarely be used at all.
So we rename it to "hasSignedSpecifier" to make it clear
that it is not about whether the type is signed.
And we rename all the other related methods in the same way.

Documentation is also added to isSigned and isUnsigned.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@44599 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c