Fix offline builds of CPAchecker
As long as all required dependencies are available locally,
this is supposed to work and usually does.
However, SpotBugs introduced a dependency that caused Ivy
to look up the newest version online.
We do not want this, so let's just fix some version of that dependency.

Also clarify in the docs that this is supposed to work.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@44600 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c