Fix CI job javascript-e2e-tests
Protractor's downloading of chromedriver is broken
(https://github.com/angular/webdriver-manager/issues/523),
so we add a workaround using a different download tool.

We could do this by changing the CI job definition,
but then we would have to sync this change to every branch.
If we instead change only the Docker image
and hack the npx command (that gets executed by the current CI job)
to do what we need, it will immediately start working again
across all branches, and it will be easy to remove again if necessary.
So while this hack is somewhat ugly, it seems preferable.

Cf. #1141

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@44685 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c