Add automaton expression:check if line is reached
This is different form CheckCoversLines. For now we do not bother
with checking a set of lines While CheckCoversLines checks whether
the CFAEdge covers one of the lines, CheckReachesLine checks whether
we reach a CFANode with the CFAEdge from which there is a CFAEdge
that covers that line. This is useful if one wants to determine
whether one is at a point before a statement in the line to be
reached is executed.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@44769 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c