Make CPAchecker's temporary variables const where possible.
Some of our temporary variables have no initializer and are assigned
later, but most of the temporary variables have an initializer and are
used only to read the value (e.g., to store the previous value in an x++
expression). Let's make those const because it better conveys the
meaning and it might be useful for some analyses.
I did not find any case where an initialized temporary variable
would be modified later on.

This implements a TODO that is 9 years and 11 months old. :-)

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@44778 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c