Simplify manual field-offset computations
Some code for raw pointer access to fields uses calculations of the form
(int) (& ((struct s *)0)->field)
to compute the offset of field in struct s.
Several of the LDV tasks even contain code like this inside array declarations
(presumably to declare arrays inside structs inside unions
that match the length of other (nested) members).
Some analyses (like the predicate analyses) work better if the array
size is statically known and has the form of a simple number,
so we can simplify this for them.

Furthermore, this is a requirement for #1146
because for this issue we need to simplify all array-length expressions
that occur in global variables.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@44779 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c