Add temporary variables for (complex) array-length expressions
Part of #1146.

The reason to do this is that it is easier and less error-prone
to handle code like "int a[size]; sizeof(a)"
if the analysis can rely on "size" being either a constant or a const variable.
Some existing analyses even treat other array lengths incorrectly.

This commit solves the problem for regular array declarations,
but not for arrays in function parameters,
where adding a temporary variable would be very complicated.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@44783 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c