Fix recursive use of subclasses of BaseSizeofVisitor for structs
Previously calling MachineModel#getSizeof(CType, BaseSizeofVisitor)
would use the passed visitor, but as soon as the recursive handling of
types would encounter a struct type, the sizes of the members of the
struct would be calculated using the standard BaseSizeofVisitor,
not the one that the user passed in.
Now we make sure to keep using the current visitor,
so recursive calculation of types with a custom visitor works as
expected.

For unions the code was already correct before.

The only caller for which this might have an effect is currently
SMGCPAExpressionEvaluator#getBitSizeof() of the cpa.smg2 package.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@44787 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c