Use different key for yml witness export
Not sure how Jackson determines the name of the attribute that is exported,
it seems I have to both change the getter name and attribute name inside
the class, and it probably does some inference based on that if I do
not provide manual serialization hints. I checked with the tests
that this works as intended.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@44801 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c