Make sizeof handling in predicate analysis (mostly) accurate
Previously, variable-length arrays were approximated using a constant
value. Now we create a formula for the correct expression.
Fixes #1067.

However, there is one remaining case that is too tricky to solve right
now, which is the one described in
https://gitlab.com/sosy-lab/software/cpachecker/-/issues/1146#note_1579941122
for variable-length arrays in function parameters.
If this occurs and the variables from the array length
are modified in the function, the value can be wrong.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@44807 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c