Restructure type-size handling in predicate analysis
This is a preparation for #1030.
So far we have one method that computes the size of a type statically,
and returns a default value if the size is not known.
But many callers actually only pass types where the size is expected to
be known, and it is unclear where the default value is actually relied on.

So now we introduce two methods, one that uses the default value,
and one that refuses to compute the size if it is not known statically.
All callers are switched to the latter except if tests have shown
that the respective caller actually needs the one with the default
value. These are surprisingly few.

If I missed some callers and this commit produces crashes,
then either solve the reliance on the default value
by using getSizeExpression() (which produces a formula for the exact value)
or switch the caller to getApproximatedSizeof().

As both introduced methods always return the same value
if they do not crash, there are no other resulting effects.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@44814 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c