Fix two-dimensional arrays for predicate analyis without pointer aliasing
Part of #1030

When accessing a two-dimensional array
we need the exact size of the inner array,
which was not working for variable-length arrays.

Given that these were the last calls to getApproximatedSizeof()
from the base predicate-analysis package (outside of pointeraliasing),
the configuration of predicate analysis without pointer aliasing
should now not rely on and default array lengths anymore.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@44815 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c