Fix crashes in InvariantsCPA from r44784 / 7eb76718
The ExpressionToFormulaVisitor insists in creating formulas for all
variables, no matter whether they are supported or not.
I don't understand the component enough to adjust that.
But we also do not want to keep using types with a size that is just
wrong for variable-length arrays, as we did before.
So lets use a dummy type with size zero.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@44845 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c