Fix offsets for two-dimensional variable-length arrays in value analysis
When accessing a two-dimensional array where the inner dimension
has variable length, we need to use that length expression
when computing the offset to use.
So far the value analysis
used the wrong value returned by the machine model.

This fixes the same problem as discussed in #1030,
just for the predicate analysis.
It is also a prerequisite for fixing #361.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@44892 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c