Fix wrong use of approximated array size in AssignmentHandler
The code does not want to handle partial assignments,
but it did so by comparing the assignment size with the approximated
array size. This means that it mistakenly would perform the assignment
if the approximated array size happens to match.
Instead we want to check whether the size is known,
and only if it is and it matches, then we do the assignment.

This improves handling of variable-length arrays
(#1030 and one-dimensional ones as well).

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@44918 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c