Avoid interpolation in predicate analysis for single-block error paths
When we analyze an error path that consists of a single block
(i.e., no loop head or other abstraction location on the path),
interpolation is useless because we would not get any non-trivial
interpolants at all. So using an interpolating solver is not necessary.
It also shouldn't hurt, but MathSAT actually fails to prove
unsatisfiability in such cases for formulas with floats
(because it does not support interpolation for floats, it just aborts).
So we better use a non-interpolating solver for single-block paths.

This gives about 250 more correct results on SV-Benchmarks
because there are many programs with floats and no loops.
But note that these programs could already be solved by BMC,
the improvement is just for predicate analysis with CEGAR
(e.g., -predicateAnalysis).

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@44946 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c